+++ lambeq/text2diagram/__init__.py
@@ -49,9 +49,9 @@ from lambeq.text2diagram.depccg_parser import DepCCGParseError, DepCCGParser
 from lambeq.text2diagram.web_parser import WebParseError, WebParser
 
 from lambeq.text2diagram.linear_reader import (LinearReader,
-                                               bag_of_words_reader,
                                                cups_reader,
-                                               spiders_reader,
                                                stairs_reader,
                                                word_sequence_reader)
+from lambeq.text2diagram.spiders_reader import (bag_of_words_reader,
+                                                spiders_reader)
 from lambeq.text2diagram.tree_reader import TreeReader, TreeReaderMode