+++ lambeq/training/tket_model.py
@@ -68,7 +68,7 @@ class TketModel(QuantumModel):
         return measured.lambdify(*self.symbols)
 
     def _randint(self, low=-1 << 63, high=(1 << 63)-1):
-        return np.random.randint(low, high)
+        return np.random.randint(low, high, dtype=np.int64)
 
     def get_diagram_output(self, diagrams: list[Diagram]) -> np.ndarray:
         """Return the prediction for each diagram using t|ket>.
@@ -106,9 +106,9 @@ class TketModel(QuantumModel):
         # discopy evals a single diagram into a single result
         # and not a list of results
         if len(diagrams) == 1:
-            result = self._normalise(tensors.array)
+            result = self._normalise_vector(tensors.array)
             return result.reshape(1, *result.shape)
-        return np.array([self._normalise(t.array) for t in tensors])
+        return np.array([self._normalise_vector(t.array) for t in tensors])
 
     def forward(self, x: list[Diagram]) -> np.ndarray:
         """Perform default forward pass of a lambeq quantum model.